--- src/Components/AppConstants.js
@@ -1,3 +1,0 @@
-const ValidImageFormats = ['jpg', 'png']
-
-export { ValidImageFormats }
